+++ src/libtriton/ast/representations/astSmtRepresentation.cpp
@@ -362,7 +362,7 @@ namespace triton {
 
       /* forall representation */
       std::ostream& AstSmtRepresentation::print(std::ostream& stream, triton::ast::ForallNode* node) {
-        triton::uint32 size = node->getChildren().size() - 1;
+        triton::usize size = node->getChildren().size() - 1;
 
         stream << "(forall (";
         for (triton::uint32 i = 0; i != size; i++) {
